cleanup exceptions in pybind11 and boost python
