+++ CMakeLists.txt
@@ -865,7 +865,6 @@ if(OPT_PYTHON AND PYTHON_EXECUTABLE)
 
   # build python bindings, either with pybind11 or Boost.Python
   set(USE_PYBIND11_PYTHON_BINDINGS ON CACHE BOOL "Build OpenRAVE's python bindings with pybind11 (ON) or Boost.Python (OFF)")
-  set(PYBIND11_PYTHON_BINDINGS_DEBUG OFF)
   add_subdirectory(python)
 endif()
