+++ python/openravepy_ext.py
@@ -48,8 +48,8 @@ class CollisionOptionsStateSaver(object):
             success = self.checker.SetCollisionOptions(self.newoptions)
             if not success and self.required:
                 self.checker.SetCollisionOptions(self.oldoptions)
-                raise openrave_exception('Failed to set options 0x%x on checker %s'%(self.newoptions,str(self.checker.GetXMLId())))
-            
+                raise ValueError('Failed to set options 0x%x on checker %s'%(self.newoptions,str(self.checker.GetXMLId())))
+    
     def __exit__(self, type, value, traceback):
         if self.oldoptions is not None:
             self.checker.SetCollisionOptions(self.oldoptions)
@@ -100,53 +100,6 @@ _registerEnumPicklers()
 import atexit
 atexit.register(openravepy_int.RaveDestroy)
 
-class openrave_exception_helper(Exception):
-    # wrap up the C++ openrave_exception
-    def __init__( self, app_error ):
-        Exception.__init__( self )
-        self._pimpl = app_error
-    def __str__( self ):
-        return str(self._pimpl)
-    def __unicode__( self ):
-        return unicode(self._pimpl)
-    def __getattribute__(self, attr):
-        my_pimpl = super(openrave_exception_helper, self).__getattribute__("_pimpl")
-        try:
-            return getattr(my_pimpl, attr)
-        except AttributeError:
-            return super(openrave_exception_helper,self).__getattribute__(attr)
-
-if openravepy_int.__pythonbinding__ != 'pybind11':
-    openrave_exception = openrave_exception_helper
-
-class std_exception(Exception):
-    """wrap up the C++ std_exception"""
-    def __init__( self, app_error ):
-        Exception.__init__( self )
-        self._pimpl = app_error
-    def __str__( self ):
-        return self._pimpl.message()
-    def __getattribute__(self, attr):
-        my_pimpl = super(std_exception, self).__getattribute__("_pimpl")
-        try:
-            return getattr(my_pimpl, attr)
-        except AttributeError:
-            return super(std_exception,self).__getattribute__(attr)
-
-class runtime_error(Exception):
-    """wrap up the C++ runtime_error"""
-    def __init__( self, app_error ):
-        Exception.__init__( self )
-        self._pimpl = app_error
-    def __str__( self ):
-        return self._pimpl.message()
-    def __getattribute__(self, attr):
-        my_pimpl = super(runtime_error, self).__getattribute__("_pimpl")
-        try:
-            return getattr(my_pimpl, attr)
-        except AttributeError:
-            return super(runtime_error,self).__getattribute__(attr)
-        
 class PlanningError(Exception):
     def __init__(self,parameter=u'', recoverySuggestions=None):
         """:param recoverySuggestions: list of unicode suggestions to fix or recover from the error
