+++ python/pybind11_bindings/CMakeLists.txt
@@ -148,7 +148,7 @@ elseif( Boost_PYTHON_FOUND AND Boost_THREAD_FOUND )
   if( HAVE_ALL_PYTHON_HEADERS )
     message(STATUS "python and boost-python found")
     message(STATUS "CMAKE_CURRENT_SOURCE_DIR = ${CMAKE_CURRENT_SOURCE_DIR}")
-    include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${PYTHON_INCLUDE_DIRS} ${OPENRAVE_INCLUDE_LOCAL_DIRS} ${OPENRAVE_CORE_INCLUDE_LOCAL_DIRS})
+    include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include ${PYTHON_INCLUDE_DIRS} ${OPENRAVE_INCLUDE_LOCAL_DIRS} ${OPENRAVE_CORE_INCLUDE_LOCAL_DIRS})
 
     # libraries to link
     set(OPENRAVEPY_LINKED_LIBRARIES "${STDC_LIBRARY} ${PYTHON_LIBRARIES} ${Boost_PYTHON_LIBRARY} ${Boost_THREAD_LIBRARY}")
@@ -188,7 +188,7 @@ elseif( Boost_PYTHON_FOUND AND Boost_THREAD_FOUND )
     if( ANN_FOUND )
       add_library(pyANN_int SHARED pyann.cpp bindings.cpp)
       include_directories(${ANN_INCLUDE_DIR})
-      target_link_libraries(pyANN_int PUBLIC ${OPENRAVEPY_LINKED_LIBRARIES} ANN PRIVATE boost_assertion_failed)
+      target_link_libraries(pyANN_int PUBLIC ${OPENRAVEPY_LINKED_LIBRARIES} libopenrave ANN PRIVATE boost_assertion_failed)
       set_target_properties(pyANN_int PROPERTIES PREFIX "" COMPILE_FLAGS "${ANN_CFLAGS}" )
       add_dependencies(pyANN_int ANN)
       if( WIN32 )
@@ -207,7 +207,7 @@ elseif( Boost_PYTHON_FOUND AND Boost_THREAD_FOUND )
     if( CONVEXDECOMPOSITION_FOUND )
       add_library(convexdecompositionpy SHARED convexdecompositionpy.cpp bindings.cpp)
       include_directories(${CONVEXDECOMPOSITION_INCLUDE_DIR})
-      target_link_libraries(convexdecompositionpy PUBLIC ${OPENRAVEPY_LINKED_LIBRARIES} convexdecomposition PRIVATE boost_assertion_failed)
+      target_link_libraries(convexdecompositionpy PUBLIC ${OPENRAVEPY_LINKED_LIBRARIES} libopenrave convexdecomposition PRIVATE boost_assertion_failed)
       set_target_properties(convexdecompositionpy PROPERTIES PREFIX "" COMPILE_FLAGS "${CONVEXDECOMPOSITION_CFLAGS}" )
       add_dependencies(convexdecompositionpy convexdecomposition)
       if( WIN32 )
