+++ python/pybind11_bindings/include/openravepy/pybind11/pybind11_bindings.h
@@ -107,14 +107,14 @@ inline py::array_t<T> toPyArrayN(const T* pvalues, const size_t N)
 {
     // one-dimension numpy array
     std::vector<npy_intp> dims {(long int)N};
-    return py::array_t<T>(dims, pvalues); 
+    return py::array_t<T>(dims, pvalues);
 }
 
 template <typename T>
 inline py::array_t<T> toPyArrayN(const T* pvalues, std::vector<npy_intp>& dims)
 {
     // n-dimension numpy array
-    return py::array_t<T>(dims, pvalues); 
+    return py::array_t<T>(dims, pvalues);
 }
 
 template <typename T>
@@ -144,6 +144,108 @@ inline py::array_t<T> toPyArray(const boost::array<T, N>& v)
 }
 
 #endif // OPENRAVE_BINDINGS_PYARRAY
+
+template <typename type>
+class PyOpenRAVEException : public py::object {
+public:
+    PyOpenRAVEException() = default;
+    PyOpenRAVEException(handle scope, const char *name, PyObject *base = PyExc_Exception, PyObject* dict = NULL) {
+        std::string full_name = scope.attr("__name__").cast<std::string>() +
+                                std::string(".") + name;
+        m_ptr = PyErr_NewException(const_cast<char *>(full_name.c_str()), base, dict);
+        if (hasattr(scope, name))
+            RAVELOG_WARN("Error during initialization: multiple incompatible "
+                         "definitions with name \"" + std::string(name) + "\"");
+        scope.attr(name) = *this;
+    }
+
+    template <typename Func, typename ... Extra>
+    PyOpenRAVEException &def(const char *name_, Func&& f, const Extra& ... extra) {
+        py::cpp_function cf(py::method_adaptor<type>(std::forward<Func>(f)), py::name(name_), py::is_method(*this),
+                            py::sibling(py::getattr(*this, name_, py::none())), extra ...);
+        attr(cf.name()) = cf;
+        return *this;
+    }
+
+    /// Uses return_value_policy::reference_internal by default
+    template <typename Getter, typename ... Extra>
+    PyOpenRAVEException &def_property_readonly(const char *name, const Getter &fget, const Extra& ... extra) {
+        return def_property_readonly(name, py::cpp_function(py::method_adaptor<type>(fget)),
+                                     py::return_value_policy::reference_internal, extra ...);
+    }
+
+    /// Uses cpp_function's return_value_policy by default
+    template <typename ... Extra>
+    PyOpenRAVEException &def_property_readonly(const char *name, const py::cpp_function &fget, const Extra& ... extra) {
+        return def_property(name, fget, py::cpp_function(), extra ...);
+    }
+
+    /// Uses return_value_policy::reference_internal by default
+    template <typename Getter, typename Setter, typename ... Extra>
+    PyOpenRAVEException &def_property(const char *name, const Getter &fget, const Setter &fset, const Extra& ... extra) {
+        return def_property(name, fget, py::cpp_function(py::method_adaptor<type>(fset)), extra ...);
+    }
+    template <typename Getter, typename ... Extra>
+    PyOpenRAVEException &def_property(const char *name, const Getter &fget, const py::cpp_function &fset, const Extra& ... extra) {
+        return def_property(name, py::cpp_function(py::method_adaptor<type>(fget)), fset,
+                            py::return_value_policy::reference_internal, extra ...);
+    }
+
+    /// Uses cpp_function's return_value_policy by default
+    template <typename ... Extra>
+    PyOpenRAVEException &def_property(const char *name, const py::cpp_function &fget, const py::cpp_function &fset, const Extra& ... extra) {
+        return def_property_static(name, fget, fset, py::is_method(*this), extra ...);
+    }
+
+    /// Uses return_value_policy::reference by default
+    template <typename Getter, typename ... Extra>
+    PyOpenRAVEException &def_property_static(const char *name, const Getter &fget, const py::cpp_function &fset, const Extra& ... extra) {
+        return def_property_static(name, py::cpp_function(fget), fset, py::return_value_policy::reference, extra ...);
+    }
+
+    /// Uses cpp_function's return_value_policy by default
+    template <typename ... Extra>
+    PyOpenRAVEException &def_property_static(const char *name, const py::cpp_function &fget, const py::cpp_function &fset, const Extra& ... extra) {
+        auto rec_fget = get_function_record(fget), rec_fset = get_function_record(fset);
+        char *doc_prev = rec_fget->doc; /* 'extra' field may include a property-specific documentation string */
+        py::detail::process_attributes<Extra ...>::init(extra ..., rec_fget);
+        if (rec_fget->doc && rec_fget->doc != doc_prev) {
+            free(doc_prev);
+            rec_fget->doc = strdup(rec_fget->doc);
+        }
+        if (rec_fset) {
+            doc_prev = rec_fset->doc;
+            py::detail::process_attributes<Extra ...>::init(extra ..., rec_fset);
+            if (rec_fset->doc && rec_fset->doc != doc_prev) {
+                free(doc_prev);
+                rec_fset->doc = strdup(rec_fset->doc);
+            }
+        }
+        def_property_static_impl(name, fget, fset, rec_fget);
+        return *this;
+    }
+
+private:
+    static py::detail::function_record *get_function_record(handle h) {
+        h = py::detail::get_function(h);
+        return h ? (py::detail::function_record *) py::reinterpret_borrow<py::capsule>(PyCFunction_GET_SELF(h.ptr()))
+               : nullptr;
+    }
+
+    void def_property_static_impl(const char *name,
+                                  py::handle fget, py::handle fset,
+                                  py::detail::function_record *rec_fget) {
+        const auto is_static = !(rec_fget->is_method && rec_fget->scope);
+        const auto has_doc = rec_fget->doc && pybind11::options::show_user_defined_docstrings();
+
+        auto property = handle((PyObject *) (is_static ? py::detail::get_internals().static_property_type
+                                             : &PyProperty_Type));
+        attr(name) = property(fget.ptr() ? fget : py::none(),
+                              fset.ptr() ? fset : py::none(),
+                              /*deleter*/ py::none(),
+                              pybind11::str(has_doc ? rec_fget->doc : ""));
+    }
+};
 
 } // end namespace openravepy
